email-automation/requirements.txt
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
gspread==5.11.3
email-validator==2.0.0
python-dotenv==1.0.0
